Deactivated the FAST_FCOLL_TABLES options for ACGs by default
ACGs are the ones that suffer the most from loss of precision at low z from the FAST_FCOLL_TABLES approximation, so we deactivate the approximation for them by default. It can be turned on with the FLAG_FAST_ATOMIC flag in Constants.h for advanced users. This does not result in a significant speed reduction for boxes with HII_DIM>=100