Add hooks and purge IC data to conserve memory.
This adds the `hooks` keyword to each of the lower-level and higher-level functions.
This makes it possible to do essentially *anything* after computing any of the outputs,
including writing it. It also means that the user can specifically choose which boxes
to write out.

It also adds routines to prepare ICs for perturb and spin_temp, i.e. dropping all the boxes
it doesn't need, and retaining all the ones it does. Only does this if the box is written
to cache, so we know we can get the dropped data back again if required.