Fix Meson build configuration for CUDA support
Remove unsupported 'required' keyword argument from meson.get_compiler()
call. The 'required' parameter is not available in Meson 1.9.0, and
since the CUDA language is only added when both USE_CUDA=TRUE and nvcc
is found, the compiler will always be available when this code is reached.

This fixes the build error:
  meson.get_compiler got unknown keyword arguments 'required'

Signed-off-by: Angus Gray-Weale <gusgw@gusgw.met>